+++ source/stemmer_porter.h
@@ -14,7 +14,7 @@ class ANT_stemmer_porter : public ANT_stemmer, public ANT_porter
 {
 public:
 	ANT_stemmer_porter(ANT_search_engine *search_engine) : ANT_stemmer(search_engine), ANT_porter() {}
-	virtual long stem(char *term, char *destination) { return ANT_porter::stem(term, destination); }
+	virtual size_t stem(char *term, char *destination) { return ANT_porter::stem(term, destination); }
 	
 } ;
